factor out CsvDownload.svelte
