+++ ui/components/ECharts.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
   import {init} from 'echarts'
   import {onDestroy, onMount, untrack} from 'svelte'
-  import {rowsToCsv} from '../../lang/csv.ts'
   import ErrorDisplay from '../internal/ErrorDisplay.svelte'
   import {componentLogger, logExtraProps} from '../internal/telemetry.ts'
   import {enrich, horizontalBarCount} from '../component-utilities/enrich.ts'
   import type {EChartsConfig, NormalConfig, QueryResult} from '../component-utilities/types.ts'
   import '../component-utilities/theme.ts'
+  import CsvDownload from './CsvDownload.svelte'
   import Skeleton from './Skeleton.svelte'
 
   interface Props {
@@ -50,7 +50,6 @@
   function handleResults (res: QueryResult) {
     chartError = null
     loaded = res
-    registerChartExport()
     if (res?.error) chartLogger.error(res.error, {...res.error, componentId: displayId})
   }
 
@@ -68,7 +67,6 @@
       }
     } else {
       loaded = data
-      registerChartExport()
     }
   })
 
@@ -76,7 +74,6 @@
     resizeObserver?.disconnect()
     resizeObserver = null
     window.$GRAPHENE.unsubscribe(handleResults)
-    delete window.$GRAPHENE.chartExports?.[displayId]
     destroyChart()
   })
 
@@ -177,43 +174,11 @@
     return dim
   }
 
-  function registerChartExport() {
-    if (!loaded || loaded.error) return
-    window.$GRAPHENE.chartExports ||= {}
-    window.$GRAPHENE.chartExports[displayId] = {rows: loaded.rows || [], fields: loaded.fields || []}
-  }
-
-  function downloadCsv() {
-    if (!loaded || loaded.error) return
-
-    let csv = rowsToCsv(loaded.rows || [], loaded.fields || [])
-    let blob = new Blob([csv], {type: 'text/csv;charset=utf-8'})
-    let url = URL.createObjectURL(blob)
-    let link = document.createElement('a')
-    link.href = url
-    link.download = `${csvFileName(chartTitle || displayId || 'graphene-chart')}.csv`
-    document.body.appendChild(link)
-    link.click()
-    link.remove()
-    URL.revokeObjectURL(url)
-  }
-
-  function csvFileName(value: string) {
-    let normalized = value.toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/^-+|-+$/g, '')
-    return normalized || 'graphene-chart'
-  }
-
 </script>
 
 <div class="echarts" bind:this={node} style={chartSizeStyle} data-component-id={mountedComponentId} data-chart-title={chartTitle}>
   {#if loaded && !loaded.error && !chartError}
-    <button class="csv-download" type="button" aria-label="Download chart data as CSV" title="Download chart data as CSV" onclick={downloadCsv}>
-      <svg viewBox="0 0 24 24" aria-hidden="true">
-        <path d="M12 3v12" />
-        <path d="m7 10 5 5 5-5" />
-        <path d="M5 21h14" />
-      </svg>
-    </button>
+    <CsvDownload data={loaded} exportId={displayId} title={chartTitle} />
   {/if}
   {#if loaded?.error || chartError}
     <ErrorDisplay error={loaded?.error || chartError} />
@@ -227,52 +192,6 @@
 <style>
   .echarts {
     position: relative;
-  }
-
-  .csv-download {
-    position: absolute;
-    top: -0.375rem;
-    right: 1rem;
-    z-index: 2;
-    display: grid;
-    place-items: center;
-    width: 1.75rem;
-    height: 1.75rem;
-    padding: 0;
-    color: #6b7280;
-    background: transparent;
-    border: 0;
-    border-radius: 0.375rem;
-    cursor: pointer;
-    opacity: 0;
-    transition: opacity 120ms ease, color 120ms ease;
-  }
-
-  .echarts:hover .csv-download,
-  .echarts:focus-within .csv-download,
-  .csv-download:hover,
-  .csv-download:focus-visible {
-    opacity: 1;
-  }
-
-  .csv-download:hover,
-  .csv-download:focus-visible {
-    color: #111827;
-  }
-
-  .csv-download:focus-visible {
-    outline: 2px solid #2563eb;
-    outline-offset: 2px;
-  }
-
-  .csv-download svg {
-    width: 1rem;
-    height: 1rem;
-    fill: none;
-    stroke: currentColor;
-    stroke-width: 2;
-    stroke-linecap: round;
-    stroke-linejoin: round;
   }
 
   .empty-chart {
